pyproject.toml
[tool.poetry]
name = "assistpy"
version = "0.1.0"
description = "Assistant made using Python"
authors = ["Devansh3712 <devanshamity@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.25.1"
python-vlc = "^3.0.11115"
pafy = "^0.5.5"
wikipedia = "^1.4.0"
pyttsx3 = "^2.90"
SpeechRecognition = "^3.8.1"
mysql-connector-python = "^8.0.23"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
